+++ package.json
@@ -1,5 +1,5 @@
 {
-    "name": "synthswap",
+    "name": "@kwenta/synthswap",
     "version": "1.0.0",
     "main": "utils/index.js",
     "scripts": {
